Separate door abstraction
